monitor.c: avoid possibility of truncated results from snprintf
mon_start() in monitor.c may cause snprintf to return a truncated
result due to being passed an output buffer which is too small to
safely hold the maximal length result when setting up the value of
new_name to be passed to task_name_set().

This is probably harmless as a task name will be truncated anyway if
it exceeds the limits of the kernel (ie 15 chars + 1 null) however it
causes the following compiler warnings with gcc 8:

 | src/lib/common/src/monitor.c: In function 'mon_start':
 | src/lib/common/src/monitor.c:155:53: error: '.master' directive output may be truncated writing 7 bytes into a region of size between 1 and 16 [-Werror=format-truncation=]
 |              snprintf(new_name, sizeof(new_name), "%s.master", old_name);
 |                                                      ^~~~~~~
 | src/lib/common/src/monitor.c:155:13: note: 'snprintf' output between 8 and 23 bytes into a destination of size 16
 |              snprintf(new_name, sizeof(new_name), "%s.master", old_name);
 |              ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 | src/lib/common/src/monitor.c:140:57: error: '.slave' directive output may be truncated writing 6 bytes into a region of size between 1 and 16 [-Werror=format-truncation=]
 |                  snprintf(new_name, sizeof(new_name), "%s.slave", old_name);
 |                                                          ^~~~~~
 | src/lib/common/src/monitor.c:140:17: note: 'snprintf' output between 7 and 22 bytes into a destination of size 16
 |                  snprintf(new_name, sizeof(new_name), "%s.slave", old_name);
 |                  ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

The fix is in two parts: firstly increase the size of the new_name
buffer to avoid the chance of truncation when the ".master" or
".slave" suffix is appended and also remove the strlen() check from
within task_name_set() - it's not necessary to skip the call to
prctl(PR_SET_NAME, ...) if the new name is too long as the name will
be automatically truncated within prctl(PR_SET_NAME, ...) as
required.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>