SM: allow to avoid of using reporting interval delay
Make it possible to avoid of using reporting interval delay when sending
statistics.
The data will be sent by the SM every sampling_interval seconds
to the SM-QM queue if the reporting_interval is set to -1.
Previously the minimal reporting_interval delay was 1 second.
It can be useful when it is needed to collect fresh WiFi/device
statistics in real-time.

Signed-off-by: Comcast-Serhiy-Chumak <serhiy_chumak@comcast.com>