wm2: ignore some radio config parameters with no VIFs
The existing code tried to ignore the channel during config and state
comparison when no VIFs are configured. In this situation, radio config
and state would never converge since the target cannot set a channel
without at least one VIF. However, overwriting the channel value still
reports a change if a channel exists in the config but not the state.
This causes wm2 to regularly try to reconfigure the radio to no effect.
To reproduce this, set a fixed channel in the radio config and run the
vif_clean FUT script.

Refactor the code to clear channel_exists in the radio config since
non-existing values are ignored during comparison. While here, add the
same treatment for ht_mode and hw_mode which similarly require a VIF
to be correctly reported in the state by the target implementation.

Signed-off-by: Stefan Schake <stefan.schake@devolo.de>