Added kconfig includes for service provider layer
kconfig includes which allow for service provider menu and managers
are missing from current master. `kconfig/Kconfig` was modified to
allow for a service provider menu similar to platform with the
exception of no default service provider. And
`kconfig/Kconfig.managers` was modified to include manager kconfig
defined by the service provider.

Since service provider is a many to many or none relationship with
platform it is best to have the includes in the core opensync portion.